Remove unused @Parcelize for Sprint, fix order type for Swimlane (Long)

+++ app/src/main/java/io/eugenethedev/taigamobile/domain/entities/Swimlane.kt
@@ -3,5 +3,5 @@ package io.eugenethedev.taigamobile.domain.entities
 data class Swimlane(
     val id: Long,
     val name: String,
-    val order: Int
+    val order: Long
 )
